+++ examples/mlperf/unet3d/training.py
@@ -1,13 +1,14 @@
 import random
 import time
+import logging
 
 import numpy as np
 import torch
 from tqdm import tqdm
 
-from examples.mlperf.metrics import dice_ce_loss
-from examples.mlperf.unet3d.inference import evaluate
+from examples.mlperf.metrics import dice_ce_loss, get_dice_score_np
 from extra.datasets import kits19
+from extra.datasets.kits19 import sliding_window_inference
 from extra.lr_scheduler import MultiStepLR
 from extra.training import lr_warmup
 from tinygrad.helpers import dtypes, getenv
@@ -22,18 +23,35 @@ from examples.mlperf.unet3d.data_loader import get_data_loaders
 from examples.mlperf.unet3d.flags import Flags
 from models.unet3d import UNet3D
 
+def evaluate(flags, model, loader, score_fn=get_dice_score_np, epoch=0):
+    s, i = 0, 0
+    for i, batch in enumerate(tqdm(loader, disable=not flags.verbose)):
+      image, label = batch
+      image, label = image.numpy(), label.numpy()
+      output, label = sliding_window_inference(model, image, label)
+      label = label.squeeze(axis=1)
+      score = score_fn(output, label).mean()
+      s += score
+      del output, label
+
+    val_dice_score = s / (i+1)
+
+    return {"epoch": epoch, "mean_dice": val_dice_score}
+
 def train(flags, model:UNet3D, train_loader, val_loader, loss_fn):
+  time_str = time.strftime("%Y_%m_%d-%H_%M_%S")
+  logging.basicConfig(filename=f'train_log_{time_str}.log', level=logging.INFO)
+  logging.info('START')
+
   is_successful, diverged = False, False
   optimizer = optim.SGD(get_parameters(model), lr=flags.learning_rate, momentum=flags.momentum, weight_decay=flags.weight_decay)
-  # scaler = GradScaler() # scalar is only needed when doing mixed precision. The default args have this disabled.
   if flags.lr_decay_epochs:
-    scheduler = MultiStepLR(optimizer, milestones=flags.lr_decay_epochs, gamma=flags.lr_decay_factor)
+    scheduler = MultiStepLR(optimizer, milestones=flags.lr_decay_epochs, gamma=flags.lr_decay_factor) # not used at the moment
   next_eval_at = flags.start_eval_at
 
-  jit_model = TinyJit(lambda x: model(x).realize()) if getenv("JIT") else model # todo this might be nicer. Such that it can also be used for evaluate
+  jit_model = TinyJit(lambda x: model(x).realize()) if getenv("JIT") else model
 
   def training_step(model_output, label, lr):
-    optimizer.zero_grad()
     print("No jit (yet)")
     optimizer.lr = lr
     loss_value = loss_fn(model_output, label)
@@ -42,25 +60,13 @@ def train(flags, model:UNet3D, train_loader, val_loader, loss_fn):
     optimizer.step()
     return loss_value.realize()
   training_step_fn = TinyJit(training_step) if getenv("JIT") else training_step
-  if getenv("OVERFIT"):
-    loader = [(0, next(iter(train_loader)))]
 
   for epoch in range(1, flags.max_epochs + 1):
-    Tensor.no_grad = False
-    print('epoch', epoch)
     cumulative_loss = []
     if epoch <= flags.lr_warmup_epochs:
       lr_warmup(optimizer, flags.init_learning_rate, flags.learning_rate, epoch, flags.lr_warmup_epochs)
-    loss_value = None
-    if not getenv("OVERFIT"):
-      loader = enumerate(tqdm(train_loader, disable=not flags.verbose))
-      # print('len(loader)', len(loader))
-      # loader = loader[:4]
-    # 1 EPOCH with FILTERS="8 64 128 256 320" FP16=0 With 24 images currently takes 31 seconds. ~1.3 seconds per step. LOSS is actually going down!! To around 0.3 - 0.5
-    # dice score is now also at 0.3-0.4 mean: 0.298
     start_time_epoch = time.time() # for 19 steps its currently ~5 seconds.
-    for iteration, batch in loader:
-      # print('optimizer.lr', optimizer.lr.numpy())
+    for iteration, batch in enumerate(tqdm(train_loader, disable=not flags.verbose)):
       image, label = batch
 
       dtype_img = dtypes.half if getenv("FP16") else dtypes.float
@@ -73,21 +79,21 @@ def train(flags, model:UNet3D, train_loader, val_loader, loss_fn):
       del image
       loss_value = training_step_fn(output, label, optimizer.lr)
       del output, label
-      # print('grad', loss_value.grad.numpy())
+      loss_value = loss_value.numpy()
       cumulative_loss.append(loss_value)
-      print('loss_value', loss_value.numpy())
+
       if flags.lr_decay_epochs:
         scheduler.step()
+    logging.info(f'loss_value epoch {epoch} {sum(cumulative_loss) / len(cumulative_loss)}')
 
     if epoch == next_eval_at and getenv("EVAL", 1):
       next_eval_at += flags.evaluate_every
-      Tensor.no_grad = True
       dtype_img = dtypes.half if getenv("FP16") else dtypes.float
 
       eval_model = lambda x : model(Tensor(x, dtype=dtype_img)).numpy() # todo maybe change
       eval_metrics = evaluate(flags, eval_model, val_loader, epoch=epoch)
-      eval_metrics["train_loss"] = (sum(cumulative_loss) / len(cumulative_loss)).numpy().item()
-
+      eval_metrics["train_loss"] = sum(cumulative_loss) / len(cumulative_loss)
+      logging.info(eval_metrics)
       Tensor.training = True
       print('eval_metrics', [(k, f"{m:.7f}") for k,m in eval_metrics.items()])
       if eval_metrics["mean_dice"] >= flags.quality_threshold:
@@ -104,12 +110,9 @@ def train(flags, model:UNet3D, train_loader, val_loader, loss_fn):
 if __name__ == "__main__":
   print('Device', Device.DEFAULT)
   # batch_size 2 is default: https://github.com/mlcommons/training/blob/00f04c57d589721aabce4618922780d29f73cf4e/image_segmentation/pytorch/runtime/arguments.py
-  # this is the real starting script: https://github.com/mlcommons/training/blob/00f04c57d589721aabce4618922780d29f73cf4e/image_segmentation/pytorch/run_and_time.sh
-  # todo check batch size?
-  print(kits19.BASEDIR)
   flags = Flags(batch_size=getenv("BATCH", 1), verbose=True, data_dir=getenv("DATA_DIR", kits19.BASEDIR))#os.environ["KITS19_DATA_DIR"])
   # flags = Flags(batch_size=getenv("BATCH", 1), verbose=True, data_dir=getenv("DATA_DIR", '/home/gijs/code_projects/tinyrad/extra/datasets/kits19/data'))#os.environ["KITS19_DATA_DIR"])
-  flags.num_workers = 0 # for debugging
+  flags.num_workers = 0 if getenv("SPEED", 2) > 0 else 1
   seed = flags.seed
   flags.evaluate_every = getenv("EVAL_STEPS", 100) # todo
   flags.start_eval_at = getenv("EVAL_START", 1) # todo
